+++ packages/flexysearch-react/src/state/state.ts
@@ -1,0 +1,8 @@
+import { FlexysearchHookState } from './../../types/index';
+
+const initialState: FlexysearchHookState = {
+    data: [],
+    filtered_data: [],
+    search_term: '',
+}
+export default initialState