+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2023-03-21-flexysearch-react
+
+### Changes
+
+- Improve global search with pattern regexp
+
+## [0.1.0] - 2023-03-21-flexysearch-react
+
+### Added
+
+- Start project and testing
+
 ## [1.0.0] - 2022-10-09
 
 ### Added